sequelize transaction

+++ server/controller/expenseController.js
@@ -1,3 +1,4 @@
+const sequelize = require('../database/configDatabase')
 const Expense = require('../model/Expense')
 const User = require('../model/User')
 
@@ -5,18 +6,23 @@ const User = require('../model/User')
 
 
 const createExpense = async(req,res) => {
+
+    const t = await sequelize.transaction()
     try {
         const {amount,description,category} = req.body  
 
-        const newExpense = await Expense.create({amount,description,category, userId: req.userId})
+        const newExpense = await Expense.create({amount,description,category, userId: req.userId},{t})
 
 
         const user = await User.findByPk(req.userId)
 
         const updateTotalExpense = user.totalexpenses + parseInt(amount)
-        await user.update({totalexpenses: updateTotalExpense})
+        await user.update({totalexpenses: updateTotalExpense},{t})
+
+      await  t.commit()
         res.status(201).json(newExpense)
      }catch(err) {
+        await t.rollback()
         console.log(err)
         res.status(500).json({message:'internal server error'})
      }
@@ -35,7 +41,20 @@ const getExpense = async(req,res) => {
 const deleteExpense = async(req,res) => {
     try {
         const {id} =req.params
+
+        const deletedExpense = await Expense.findByPk(id)
+
+        if (!deletedExpense) {
+            return res.status(404).json({ message: 'Expense not found' });
+        }
+
+        const deletedAmount = deletedExpense.amount
         await Expense.destroy({where : {id,userId:req.userId}})
+
+        const user = await User.findByPk(req.userId)
+
+        const updatedTotalExpenses = user.totalexpenses - deletedAmount
+        await user.update({totalexpenses : updatedTotalExpenses})
         res.status(204).end()
     }catch(err){
         console.log(err)
